+++ src/Kyrodan.HiDrive.Tests/FileTests.cs
@@ -1,0 +1,15 @@
+﻿using System.Threading.Tasks;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Kyrodan.HiDrive.Tests
+{
+    [TestClass]
+    public class FileTests : BaseTest
+    {
+        [TestMethod]
+        public async Task Get()
+        {
+            Assert.IsTrue(true);
+        }
+    }
+}